ref(swaps): move `SwapStep` to `components/swap/types`
