+++ components/swap/types.ts
@@ -1,0 +1,7 @@
+export enum SwapStep {
+  COUNTERPARTY,
+  DESIRED,
+  OFFERED,
+  REVIEW,
+  CREATED,
+}
